+++ expr/functions/cairo/png/png.go
@@ -10,7 +10,7 @@ import (
 
 const HaveGraphSupport = false
 
-func EvalExprGraph(e parser.Expr, from, until int32, values map[parser.MetricRequest][]*types.MetricData) ([]*types.MetricData, error) {
+func EvalExprGraph(e parser.Expr, from, until int64, values map[parser.MetricRequest][]*types.MetricData) ([]*types.MetricData, error) {
 	return nil, nil
 }
