+++ expr/functions/reduce/function.go
@@ -27,7 +27,7 @@ func New(configFile string) []interfaces.FunctionMetadata {
 	return res
 }
 
-func (f *reduce) Do(e parser.Expr, from, until int32, values map[parser.MetricRequest][]*types.MetricData) ([]*types.MetricData, error) {
+func (f *reduce) Do(e parser.Expr, from, until int64, values map[parser.MetricRequest][]*types.MetricData) ([]*types.MetricData, error) {
 	const matchersStartIndex = 3
 
 	if len(e.Args()) < matchersStartIndex+1 {
