+++ expr/helper/align.go
@@ -6,13 +6,13 @@ import (
 )
 
 // GetBuckets returns amount buckets for timeSeries (defined with startTime, stopTime and step (bucket) size.
-func GetBuckets(start, stop, bucketSize int32) int32 {
-	return int32(math.Ceil(float64(stop-start) / float64(bucketSize)))
+func GetBuckets(start, stop, bucketSize int64) int64 {
+	return int64(math.Ceil(float64(stop-start) / float64(bucketSize)))
 }
 
 // AlignStartToInterval aligns start of serie to interval
-func AlignStartToInterval(start, stop, bucketSize int32) int32 {
-	for _, v := range []int32{86400, 3600, 60} {
+func AlignStartToInterval(start, stop, bucketSize int64) int64 {
+	for _, v := range []int64{86400, 3600, 60} {
 		if bucketSize >= v {
 			start -= start % v
 			break
@@ -23,9 +23,9 @@ func AlignStartToInterval(start, stop, bucketSize int32) int32 {
 }
 
 // AlignToBucketSize aligns start and stop of serie to specified bucket (step) size
-func AlignToBucketSize(start, stop, bucketSize int32) (int32, int32) {
-	start = int32(time.Unix(int64(start), 0).Truncate(time.Duration(bucketSize) * time.Second).Unix())
-	newStop := int32(time.Unix(int64(stop), 0).Truncate(time.Duration(bucketSize) * time.Second).Unix())
+func AlignToBucketSize(start, stop, bucketSize int64) (int64, int64) {
+	start = int64(time.Unix(int64(start), 0).Truncate(time.Duration(bucketSize) * time.Second).Unix())
+	newStop := int64(time.Unix(int64(stop), 0).Truncate(time.Duration(bucketSize) * time.Second).Unix())
 
 	// check if a partial bucket is needed
 	if stop != newStop {
